+++ R/lazyframe-frame.R
@@ -764,8 +764,8 @@ lazyframe__with_columns_seq <- function(...) {
 
 #' Remove columns
 #'
-#' @param ... <[`dynamic-dots`][rlang::dyn-dots]> Names of the columns that
-#' should be removed. Accepts column selector input.
+#' @param ... <[`dynamic-dots`][rlang::dyn-dots]> Column names or [selectors][polars_selector]
+#'   that should be removed.
 #' @param strict Validate that all column names exist in the current schema,
 #' and throw an exception if any do not.
 #'
@@ -1107,9 +1107,9 @@ lazyframe__reverse <- function() {
 #'
 #' The original order of the remaining rows is preserved.
 #'
-#' @param ... <[`dynamic-dots`][rlang::dyn-dots]> Column name(s) for which null
-#' values are considered. If empty (default), use all columns.
-#'
+#' @param ... <[`dynamic-dots`][rlang::dyn-dots]> Column names or [selectors][polars_selector]
+#'   for which are considered. If empty (default), use all columns
+#'   (same as specifying with the selector [`cs$all()`][cs__all]).
 #' @inherit as_polars_lf return
 #' @examples
 #' lf <- pl$LazyFrame(
@@ -1131,7 +1131,7 @@ lazyframe__drop_nulls <- function(...) {
     subset <- if (...length() == 0L) {
       NULL
     } else {
-      parse_into_selector(..., .strict = TRUE)$`_rselector`
+      parse_into_selector(...)$`_rselector`
     }
     self$`_ldf`$drop_nulls(subset)
   })
@@ -1141,11 +1141,8 @@ lazyframe__drop_nulls <- function(...) {
 #'
 #' The original order of the remaining rows is preserved.
 #'
-#' @param ... <[`dynamic-dots`][rlang::dyn-dots]> Column name(s) for which null
-#' values are considered. If empty (default), use all columns (note that only
-#' floating-point columns can contain `NaN`s).
-#'
 #' @inherit as_polars_lf return
+#' @inheritParams lazyframe__drop_nulls
 #' @examples
 #' lf <- pl$LazyFrame(
 #'   foo = c(1, NaN, 2.5),
@@ -1154,7 +1151,7 @@ lazyframe__drop_nulls <- function(...) {
 #' )
 #'
 #' # The default behavior of this method is to drop rows where any single value
-#' # of the row is null.
+#' # of the row is NaN.
 #' lf$drop_nans()$collect()
 #'
 #' # This behaviour can be constrained to consider only a subset of columns, as
@@ -1175,18 +1172,16 @@ lazyframe__drop_nans <- function(...) {
     subset <- if (...length() == 0L) {
       NULL
     } else {
-      parse_into_selector(..., .strict = TRUE)$`_rselector`
+      parse_into_selector(...)$`_rselector`
     }
     self$`_ldf`$drop_nans(subset)
   })
 }
 
-# TODO: @2.0 replace subset to dyn-dots and rename all arguments
+# TODO: @2.0 remove subset
 #' Drop duplicate rows
 #'
-#' @inheritParams rlang::args_dots_empty
-#' @param subset Column name(s) or selector(s), to consider when identifying
-#' duplicate rows. If `NULL` (default), use all columns.
+#' @inheritParams lazyframe__drop_nulls
 #' @param keep Which of the duplicate rows to keep. Must be one of:
 #' * `"any"`: does not give any guarantee of which row is kept. This allows
 #'   more optimizations.
@@ -1196,6 +1191,7 @@ lazyframe__drop_nans <- function(...) {
 #' @param maintain_order Keep the same order as the original data. This is
 #' more expensive to compute. Setting this to `TRUE` blocks the possibility to
 #' run on the streaming engine.
+#' @param subset `r lifecycle::badge("deprecated")` Replaced by `...` in 1.1.0.
 #'
 #' @inherit as_polars_lf return
 #' @examples
@@ -1206,22 +1202,73 @@ lazyframe__drop_nans <- function(...) {
 #' )
 #' lf$unique(maintain_order = TRUE)$collect()
 #'
-#' lf$unique(subset = c("bar", "ham"), maintain_order = TRUE)$collect()
+#' lf$unique(c("bar", "ham"), maintain_order = TRUE)$collect()
 #'
 #' lf$unique(keep = "last", maintain_order = TRUE)$collect()
 lazyframe__unique <- function(
-  subset = NULL,
   ...,
   keep = c("any", "none", "first", "last"),
-  maintain_order = FALSE
+  maintain_order = FALSE,
+  subset = deprecated()
 ) {
   wrap({
-    check_dots_empty0(...)
     keep <- arg_match0(keep, values = c("any", "none", "first", "last"))
-    if (!is.null(subset)) {
-      subset <- parse_into_selector(!!!subset)$`_rselector`
+
+    subset <- if (is_present(subset)) {
+      deprecate_warn(
+        format_warning(
+          c(
+            `!` = sprintf(
+              "The %s argument of %s is deprecated and replaced by %s as of %s 1.1.0.",
+              format_arg("subset"),
+              format_code("$unique()"),
+              format_arg("..."),
+              format_pkg("polars")
+            )
+          )
+        )
+      )
+      check_dots_empty0(...)
+
+      if (is.null(subset)) {
+        cs$all()
+      } else {
+        parse_into_selector(!!!c(subset))
+      }
+    } else if (...length() == 1L && (is.null(..1) || is.list(..1))) {
+      check_dots_unnamed()
+      deprecate_warn(
+        c(
+          `!` = format_warning(sprintf(
+            "Passing %s to the first argument of %s is deprecated as of %s 1.1.0.",
+            obj_type_friendly(..1),
+            format_code("$unique()"),
+            format_pkg("polars")
+          )),
+          i = format_warning(sprintf(
+            "Passing %s to %s instead.",
+            format_code(if (is.null(..1)) "cs$all()" else "!!!my_list"),
+            format_arg("...")
+          ))
+        )
+      )
+
+      if (is.null(..1)) {
+        cs$all()
+      } else {
+        parse_into_selector(!!!force(..1))
+      }
+    } else if (...length() == 0L) {
+      NULL
+    } else {
+      parse_into_selector(...)
     }
-    self$`_ldf`$unique(subset = subset, keep = keep, maintain_order = maintain_order)
+
+    self$`_ldf`$unique(
+      subset = subset$`_rselector`,
+      keep = keep,
+      maintain_order = maintain_order
+    )
   })
 }
 
@@ -1530,9 +1577,9 @@ lazyframe__rename <- function(..., .strict = TRUE) {
 
 #' Explode the frame to long format by exploding the given columns
 #'
-#' @param ... <[`dynamic-dots`][rlang::dyn-dots]> Column names, expressions, or
-#' a selector defining them. The underlying columns being exploded must be of
-#' the `List` or `Array` data type.
+#' @param ... <[`dynamic-dots`][rlang::dyn-dots]> Column names or [selectors][polars_selector]
+#'   defining them. The underlying columns being exploded must be of
+#'   the `List` or `Array` data type.
 #'
 #' @inherit as_polars_lf return
 #' @examples
@@ -1592,8 +1639,7 @@ lazyframe__clone <- function() {
 #' The new columns will be inserted at the location of the struct column.
 #'
 #' @param ... <[`dynamic-dots`][rlang::dyn-dots]> Name of the struct column(s)
-#' that should be unnested.
-#'
+#'   or [selectors][polars_selector]  that should be unnested.
 #' @inherit as_polars_lf return
 #' @examples
 #' lf <- pl$LazyFrame(